Add random number generator to World's state
We need a seedable random number generator. Also, we need to be able to
deterministically save and load the state of the random number generator
along with the world so that we can reproduce a simulation from a world
state file.

Note that we will need a way to pass in the seed. Also note, if
instantiating the number generator is slower, we may want to lazily
instantiate it.